requirements.txt
requests
streamlit
pandas
python-dotenv
feedparser
fpdf2
